Move SafeIndex to separate file
